+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.x.x] - xxxx-xx-xx
+### Added
+- `Image::IsLoaded()`
+- `Wave::IsLoaded()`
+- `Sound::IsLoaded()`
+- `Music::IsLoaded()`
+- Shader example
+- `raylib::DrawText()` for `std::string` support
+- `raylib::DrawTextEx()` for `std::string` support
+- `raylib::DrawTextPro()` for `std::string` support
+
+### Fixed
+- Shader constructor
+
+### Changed
+- Merged `Text.hpp` into `Functions.hpp`
+
 ## [4.0.1] - 2021-11-20
 ### Changes
 - Fix compiler warning